prototype Two PPUP: validate and store MIRs
Validate MIRs, as was previously done in DELEG, and
store the output of this validation
(`InstantaneousRewards`) in the PPUPState.

This was complicated by the fact that validation of
a MIRCert requires an AccountState. Consequently,
this required that AccountState was added to the
environment of UTXOW, UTXO and PPUP, and that the
LEDGER rule was changed for Voltaire to pass its
AccountState to UTXOW, so that UTXOW can pass it to
UTXO, which finally passes it to PPUP.