Revert "prototype Two PPUP: validate and store MIRs"
This reverts commit 6d97bc288cff2fd9b19816c8a2f252930cce3123.

This commit is reverted because a `MIRCert` inside
a `ProposalBody` cannot be validated at
submission-time, since it may not be enacted until
the next epoch. And if another MIR is enacted in
the current epoch, then enactment of the MIR in the
next epoch may fail. Thus, we must postpone
validation until enactment.